% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bParEstimate.R
\name{bParEstimate}
\alias{bParEstimate}
\title{Blinded Parameter Estimation}
\usage{
bParEstimate(dfSurv, theta, k)
}
\arguments{
\item{dfSurv}{a data.frame containing the variables:
\itemize{
  \item status: (0,1) := (no event, event)
  \item time: time to status
  \item group: grouping variable with two levels
}}

\item{theta}{a number defining the assumed hazard ratio.}

\item{k}{a number defining the balance between the two groups.
If not supplied it is calculated from the data.}
}
\value{
a data.frame containing the overall event rate as well as the event rates per group.
}
\description{
Performs a blinded parameter estimation.
}
\examples{
N <- matrix(rep(50,20),ncol=2)
df.sim <- simSurvData(N=N, lambda=-log(.7)/24, theta=.7, gamma=-log(.3)/24, distS="exponential",distC="exponential", L=20)
bParEstimate(dfSurv=df.sim, theta=.7, k=1)
}
