% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expEvents.R
\name{calc_expEvents}
\alias{calc_expEvents}
\title{calc_expEvents}
\usage{
calc_expEvents(N, theta, L = Inf, tn, lambda, sigma = 1,
  distS = "exponential", gamma, kappa = 1, distC = "exponential")
}
\arguments{
\item{N}{a two column matrix containing the size of both recruitment groups per
row, representing the number of recruited patients per group and timepoint (\code{t}).}

\item{theta}{a number greater 0 defining the assumed hazard ratio.}

\item{L}{a number defining the timepoint for administrative censoring.
If no administrative censoring is planned \code{L=Inf}. Default is L=Inf.}

\item{tn}{a numeric vector defining the recruitment time. Default is \code{0:(NROW(N)-1)}.}

\item{lambda}{a number greater 0 defining the hazard rate for the survival process of
group1. For the parametrization see \code{details}.}

\item{sigma, kappa}{a number greater 0 defining the shape parameter for the
survival- and the censor process. Only needed if distS or distC is set to weibull.
Default is 1 resulting in an exponential distribution.}

\item{distS, distC}{a character string defining the distribution of the survival-
and the censor process. Default is 'exponential'.}

\item{gamma}{a number greater 0 defining the overall hazard rate for the censor process.}
}
\value{
a vector containing the expected number of events per group.
}
\description{
!!!!!!!!!!!!! shall not be exported !!!!!!!!!!!!!!!!!!!
Calculates the expected number of events of a population given their respective
recruiting timepoint, assumed survival time of the first group, the hazard ratio
and ce censor rate. If the distributions are ....
}
\details{
With constant hazard rate \eqn{\lambda} and shape parameter \eqn{\sigma} the survival function for the weibull event process is given by
\deqn{S ( t ) = exp{ \lambda t ^ \sigma },}
with density function
\deqn{f ( t ) = \lambda \sigma t ^ { \sigma - 1 } exp{ \lambda t ^ \sigma }.}
}
\examples{
# with a recruitment time of 5 month and recruiting 30 Patients per month
# assuming a hazard ratio of .7 and a rate of two year survival of .7
# as well as a two year censor rate of .3 under exponential distribution
# and administrative censoring after 10 month.
N <- matrix(rep(50,10),ncol=2)
calc_expEvents(N=N ,theta = .7, L=10, lambda=-log(.7)/24, gamma=-log(.3)/24)
}
