% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_dfSim.R
\name{cut_dfSim}
\alias{cut_dfSim}
\title{Cuts the Data.Frame at a specific timepoint, changing status accordingly.}
\usage{
cut_dfSim(df.sim, t)
}
\arguments{
\item{df.sim}{a data.frame containing the variables:
\itemize{
  \item status: (0,1) := (no event,event)
  \item time: time to status
  \item tn: recruitment timepoint
}}

\item{t}{a number defining the timepoint at which the dataframe is cut.}
}
\value{
a data.frame with the same structure of the handed data.
}
\description{
Cuts the Data.Frame at a specific timepoint, changing status accordingly.
}
\details{
the status of events occuring at a later timepoint time > t are set to zero. Patients with a
recruitment timepoint > t are removed.
}
\examples{
N <- matrix(rep(50,10),ncol=2)
df.sim <- simSurvData(N=N, lambda=exp(.7)/12, theta=.7, gamma=exp(.3)/12, distS="exponential",distC="exponential", L=10)
df.sim_cut3 <- cut.dfSim(df.sim,3)
}
