% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hazardRate.R
\name{get_hazardRate}
\alias{get_hazardRate}
\alias{get_eventProportion}
\title{Get hazard rates}
\usage{
get_hazardRate(prop,time,theta=NULL)

get_eventProportion(hazard, time, theta = NULL)
}
\arguments{
\item{prop}{a vector of numbers greater 0 and smaller 1 defining the proportion of patients having an event before 'time'.}

\item{time}{a number defining the timeframe.}

\item{theta}{optional number defining the hazard rate. Returning(-log(1-prop) /time *  theta)}

\item{hazard}{a vector of numbers greater 0 defining the hazard rates per timeintervall of length 1.}
}
\value{
a vector defining the hazard rate per timeintevall of length 1.
}
\description{
Get hazard rates
}
\examples{
# Assuming, that 30\% of patients in the control group have an event after 24 month,
# the monthly hazard rate for the control group can be calculated by:
get_hazardRate(.3,24)

# With a assumed reduction of the hazard by 30\% in the treatment group we can calculate
# the hazard rate as follows:
get_hazardRate(.3,24,1-.3)
}
