% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dfSim.R
\name{plot.dfSim}
\alias{plot.dfSim}
\title{Plots an event based data frame}
\usage{
\method{plot}{dfSim}(dfSim, col = c("grey", "black"), ...)
}
\arguments{
\item{dfSim}{a data.frame with time status and recruitmenttime = tn.}

\item{col}{a vector of two colours. The first one defines the color of
observations without an event and the second one the color of the observations with an event.
Default is c("grey","black").}

\item{...}{additional \code{matplot()} related arguments.}
}
\value{
Generates a new plot.
}
\description{
Plots an event based data frame
}
\details{
The graph shows the observation time per patient from recruitment until administrative censoring.

Patients with an event are colored differently.

The Observations are sorted by observation time, status and group.
}
\examples{
N <- matrix(rep(50,10),ncol=2)
dfSim <- simSurvData(N=N, lambda=-log(1-.3)/24, sigma=1, theta=.7, gamma=-log(1-.2)/24, kappa=1, distS = "exponential",distC = "exponential",L=10)
plot(dfSim)

# to see the corresponding Kaplan-Meyer curve
require(survival)
fit <- survfit(Surv(time=time-tn, event = status) ~group,data = dfSim)
plot(fit)
}
\seealso{
\code{\link[graphics]{matplot}} \code{\link{simSurvData}}
}
