% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_recruitment.R
\name{plot.recruit_frame}
\alias{plot.recruit_frame}
\title{plot recruit_frame as a barplot}
\usage{
\method{plot}{recruit_frame}(data, ...)
}
\arguments{
\item{data}{a recruit_frame generated by the function \code{set_recruitment}.}

\item{...}{other \code{barplot()} parameter.}
}
\value{
opens a new barplot.
}
\description{
The standard plot function for a recruit_frame.
}
\examples{
x <- set_recruitment(targetN=c(714,357), timePoints=0:23)
plot(x)
}
