% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dfSim.R
\name{plot_dfSim}
\alias{plot_dfSim}
\title{Plots a event based data frame}
\usage{
plot_dfSim(dfSim, type = "l", lty = 1, ...)
}
\arguments{
\item{dfSim}{a data.frame with time status and recruitmenttime = tn.}

\item{type}{grafical parameter for the plottype. Default and recommended is "l".}

\item{lty}{grafical parameter for the linetype. Default and recommended is 1.}

\item{...}{}
}
\description{
Plots a event based data frame
}
