% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_fun.R
\name{prob_fun}
\alias{prob_fun}
\title{Event Distribution}
\usage{
prob_fun(r, lambda, sigma = 1, gamma, kappa = 1)
}
\arguments{
\item{r}{vector of recruitment timepoints.}

\item{lambda, sigma}{scale and shape parameters for the survival process, the latter defaulting to 1,
resulting in an exponential distribution.}

\item{gamma, kappa}{scale and shape parameters fot the censor process, the latter defaulting to 1,
resulting in an exponential distribution.}
}
\value{
\itemize{
\item Gives the distribution of the events.
\item Invalid arguments will result in return value \code{NaN}, with warning.
}
}
\description{
!!!!!!!!!!!!! shall not be exported !!!!!!!!!!!!!!!!!!!
}
\examples{
prob_fun(r=1:10,.014,sigma=1,.009,kappa=1)
prob_fun(r=c(1:10),.014,sigma=1,.009,kappa=1)
}
