% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEstimate.R
\name{rEstimate}
\alias{rEstimate}
\title{Recruitment Time Estimation}
\usage{
rEstimate(
  N,
  tn,
  M,
  tm,
  theta,
  nEvents,
  L,
  from = 1,
  lambda,
  sigma = 1,
  distS = "exponential",
  gamma,
  kappa = 1,
  distC = "exponential"
)
}
\arguments{
\item{N}{a two column matrix containing the size of both recruitment groups per
row, representing the number of recruited patients per group and timepoint (\code{t}).}

\item{tn}{a vecotr of length = \code{NROW(N)} defining the timepoint of recruiting.}

\item{M}{a two column matrix containing the size of both recruitment groups per
row, representing the number of future recruited patients per group and timepoint (\code{tm}).
If not defined it will be taken from the last row of \code{N}.}

\item{tm}{a vecotr of length = \code{NROW(M)} defining the timepoint of future recruiting.
If not defined it will be interpolated from the last two entries of \code{tn}.}

\item{theta}{a number greater 0 defining the assumed hazard ratio.}

\item{nEvents}{a number greater 0 defining the targeted number of events to achieve the
planned power.}

\item{L}{a number defining the timepoint for administrative censoring.
If no administrative censoring is planned \code{L=Inf}. Default is L=Inf.}

\item{from}{a number > 0 defining the minimum number of recruitment steps. Default is 1,
starting with the first recruited batch of patients.}

\item{lambda}{a number greater 0 defining the hazard rate for the survival process of
group1. For the parametrization see \code{details}.}

\item{sigma, kappa}{a number greater 0 defining the shape parameter for the
survival- and the censor process. Only needed if distS or distC is set to weibull.
Default is 1 resulting in an exponential distribution.}

\item{distS, distC}{a character string defining the distribution of the survival-
and the censor process. Default is 'exponential'.}

\item{gamma}{a number greater 0 defining the overall hazard rate for the censor process.}
}
\value{
a list with the minimum needed batches of recruits defined by N (\code{rec_batch}),
the minimum time for recruitment dependend on tn (\code{rec_time}) and the minimum amount of
patients (\code{rec_total_number}).
}
\description{
Calculates the minimum recruitment time and number of patients needed to achieve the targeted
number of events befor administrative study end.
}
\examples{
N <- matrix(rep(50,20),ncol=2)
tn <- 0:(NROW(N)-1)
rEstimate(N=N, tn=tn, theta=.7, nEvents=40, L=20, from = 1,
          lambda=-log(.7)/24, sigma=1, distS="exponential",
          gamma=-log(.3)/24, kappa=1, distC="exponential")
}
