% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoenfeld.R
\name{schoenfeld}
\alias{schoenfeld}
\alias{nschoenfeld}
\alias{pschoenfeld}
\title{Schnoenfeld Formula}
\usage{
nschoenfeld(theta, k = 1, alpha = 0.05, beta = 0.2,
  alternative = "one-sided")

pschoenfeld(theta, nEvent, k = 1, alpha = 0.05,
  alternative = "one-sided")
}
\arguments{
\item{theta}{vecor of hazard ratio theta.}

\item{k}{number >0. Defines the relative group sizes with the relation k:1.
If k=1, equal group sizes are assumed (Note: that the formula is symmetric around k=1).}

\item{alpha}{number between 0 and 1 defining the deserved type I error rate.
Default is 0.05.}

\item{beta}{number between 0 and 1 defining the deserved type II error rate.
Default is 0.1 resulting in a power of 1-0.1=0.9 (90\%).}

\item{alternative}{a string; either \code{'one-sided'} or \code{'two-sided'}. Default is
\code{'two-sided'}.}

\item{nEvent}{vecor of number of events.}
}
\value{
Returns the number of events needed to satisfy the given assumptions.
}
\description{
Calculates the number of events needed to achieve a targeted
power (or the power given the number of events) in a two group scenario with
the hazard ratio \code{theta}, balance between the groups \code{k} and the alpha niveau \code{alpha}.
}
\examples{
# for a assumed hazard ration of theta = .7
# in a design with a treatment to control group allocation of 2:1 resulting in k = 2
# a power of 1-beta=80\%
# and a significance niveau of alpha = .025
nschoenfeld(theta = .7,k=2,alpha=.05,beta=.1,alternative = "two-sided")

}
