% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_recruitment.R
\name{set_recruitment}
\alias{set_recruitment}
\title{generate recruitment table}
\usage{
set_recruitment(targetN, timePoints, fn = "linear", ...)
}
\arguments{
\item{targetN}{a vector of length 1 or two defining the targeted total number of recruitments.
If \code{targetN} is of length tow the total number recruitments per group is defined.
The first entry refers to the treatment group.}

\item{timePoints}{a vector defining the timepoint of each recruitment or a positive number defining the
amount of recruitments (needs \code{time} to be set).}

\item{fn}{a string defining the shape of the cummulative recruitment. Default is set to "linear".}

\item{k}{a positive number defining the k:1 treatment allocation (treatment:control). Optional, if
\code{targetN} is vector of length two.}

\item{timeLimits}{a vector of length two defining the timepoint of the first and the last recruitment.
Optional if \code{timePoints} is a vector defining each timepoint of recruitment.}
}
\value{
a \code{data.frame} containing the number of recruitments of the treament and control group \code{"T"} and \code{"C"} per timepoint \code{time}.
}
\description{
A function to build a recruitment dataframe by some general conditions e.g. targeted total number of recruitments per group
and recruitment timePoints.
}
\examples{
# Two different approches to define a recruitment of 1071 patients
# with an allokation of 2:1 (treatment:control), where 24 batches of
# recruits are wanted. The recruitment is assumed to take place at the
# beginning of each month over 24 month.

set_recruitment(targetN=c(714,357), timePoints=0:23)
set_recruitment(targetN=1071, k=2, timePoints=24, timeLimits=c(0,23))
}
