% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_studyPlan.R
\name{set_studyPlan}
\alias{set_studyPlan}
\title{Title}
\usage{
set_studyPlan(recruitment, rec_time, study_end, lambdaC, theta, gamma,
  dist_surv, dist_censor, alpha = 0.05, beta = 0.2,
  alternative = "two-sided", ...)
}
\arguments{
\item{lambdaC}{a number defining the hazard rate of the control group.}

\item{theta}{a number defining the hazard ratio. Alternatively \code{lambdaT} can be defined.}

\item{gamma}{a number defining the hazard of the censor process.}

\item{dist_surv}{a numeric defining the distribution of the survival process.}

\item{alternative}{}

\item{lambdaT}{a number defining the hazard rate of the treatment group.  Alternatively \code{theta} can be defined.}
}
\value{
a list of class \code{studyPlan}
}
\description{
Title
}
\examples{
set_studyPlan(recruitment=set_recruitment(c(2400,1200),1:20),
              rec_time = 10,
              study_end = 22,
              lambdaC = get_hazardRate(.3,24),
              theta = .7,
              gamma = get_hazardRate(.2,24),
              dist_surv = "exponential",
              dist_censor = "exponential",beta=.1)
}
