% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_BSSRed.R
\name{sim_BSSRed}
\alias{sim_BSSRed}
\title{Title}
\usage{
sim_BSSRed(N, tn, M, tm, addRecT, sim.lambdaP, sigma = sigma, theta.star,
  gamma, kappa, distS, distC, nEvents, L, nSim = 100, par.est = FALSE,
  fixed.min.studytime = TRUE, seed = 235711)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
