% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSurvData.R
\name{simSurvData}
\alias{simSurvData}
\title{Simulation of a Survival Data.Frame}
\usage{
simSurvData(N, tn, lambda, sigma, theta, gamma, kappa,
  distS = "exponential", distC = "exponential", L)
}
\arguments{
\item{N}{a two column matrix containing the size of both recruitment groups per
row, representing the number of recruited patients per group and timepoint (\code{tn}).}

\item{tn}{a vecotr of length = \code{NROW(N)} defining the timepoint of recruiting.}

\item{lambda}{a number greater 0 defining the hazard rate for the survival process of
group1. For the parametrization see \code{details}.}

\item{sigma, kappa}{a number greater 0 defining the shape parameter for the
survival- and the censor process. Only needed if distS or distC is set to weibull.
Default is 1 resulting in an exponential distribution.}

\item{gamma}{a number greater 0 defining the overall hazard rate for the censor process.}

\item{distS, distC}{a character string defining the distribution of the survival-
and the censor process. Default is 'exponential'.}

\item{L}{a number defining the timepoint for administrative censoring.
If no administrative censoring is planned \code{L=Inf}. Default is L=Inf.}
}
\value{
a data.frame including id (\code{id}), time to event (\code{time}), recruiting timepoint (\code{tn}),
grouping (\code{group}) and status (\code{status} (1 = event occured)).
}
\description{
Simulates a two group data.frame including time to event, recruiting time and status.
}
\examples{
N <- matrix(rep(50,10),ncol=2)
simSurvData(N=N, lambda=exp(.7)/12, sigma=1, theta=.7, gamma=exp(.3)/12, kappa=1, distS = "exponential",distC = "exponential",L=10)
}
