% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dfSim.R
\name{plot_dfSim}
\alias{plot_dfSim}
\title{Plots a event based data frame}
\usage{
plot_dfSim(dfSim, type = "l", lty = 1, ...)
}
\arguments{
\item{dfSim}{a data.frame with time status and recruitmenttime = tn.}

\item{type}{grafical parameter for the plottype. Default and recommended is "l".}

\item{lty}{grafical parameter for the linetype. Default and recommended is 1.}

\item{...}{additional \code{matplot()} related arguments.}
}
\value{
a plot of the supplied data.frame.
}
\description{
Plots a event based data frame
}
\examples{
N <- matrix(rep(50,10),ncol=2)
dfSim <- simSurvData(N=N, lambda=-log(1-.3)/24, sigma=1, theta=.7, gamma=-log(1-.2)/24, kappa=1, distS = "exponential",distC = "exponential",L=10)
plot_dfSim(dfSim)
}
